feat: move page to components folder
